+++ engine/src/calc/calc.c
@@ -44,7 +44,7 @@ static void update_display(calc_t *calc) {
     if (extended_display[0] == '\0') {
         strcpy(extended_display, "READY");
     }
-    int offset = strlen(extended_display) - 12;
+    int offset = (int)strlen(extended_display) - 12;
     if (strchr(extended_display, '.')) offset -= 1;
     if (offset < 0) {
         offset = 0;
