+++ engine/src/calc/log.c
@@ -21,13 +21,13 @@ void log_get_available_interval(log_t *log,
     if (last_index_out) *last_index_out = log->last_index;
 }
 
-char *log_get_entry(log_t *log, int index) {
+char *log_get_entry(log_t *log, long index) {
     return &log->mem[get_offset(log, index)];
 }
 
 void log_add_entry(log_t *log, char *entry) {
     // If log full, remove first element.
-    int count = 0;
+    long count = 0;
     if (log->first_index) {
         count = log->last_index - log->first_index + 1;
         if (count == MAX_ENTRY_COUNT) {
@@ -45,16 +45,16 @@ void log_add_entry(log_t *log, char *entry) {
     int new_offset = 0;
     if (log->last_index) {
         int last_offset = get_offset(log, log->last_index);
-        new_offset = last_offset + strlen(&log->mem[last_offset]) + 1;
+        new_offset = last_offset + (int)strlen(&log->mem[last_offset]) + 1;
         if (new_offset + strlen(entry) >= MEM_LEN) {
             new_offset = 0;
         }
     }
     int new_left = new_offset;
-    int new_right = new_offset + strlen(entry);
-    for (int index = log->first_index; index <= log->last_index; index++) {
+    int new_right = new_offset + (int)strlen(entry);
+    for (long index = log->first_index; index <= log->last_index; index++) {
         int left = get_offset(log, index);
-        int right = left + strlen(&log->mem[left]);
+        int right = left + (int)strlen(&log->mem[left]);
         if (intersects(new_left, new_right, left, right)) {
             log->first_index++;
         } else {
