+++ engine/src/calc/log.h
@@ -6,7 +6,7 @@ void log_get_available_interval(log_t *log,
                                 long *last_index_out);
 
 // Gets the entry of given 'index'.
-char *log_get_entry(log_t *log, int index);
+char *log_get_entry(log_t *log, long index);
 
 // Adds a new entry to the log.
 void log_add_entry(log_t *log, char *entry);
